New Function|Custom Export|Title of the Figure|Improved Plot|Axis Scale
Added a new function and changed the export feature so it opens a separate window to choose the directory, file name, and format from the available options. The figure title resets to 'СумДУ' until a plot is present. Improved the plot function, which can now plot multiple lines. Renamed the file 'FeCr_phase_graph.py' to 'physics_graphs.py' because it will now be a collection of different functions.